comsys-hackathon-taskB/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.2
albumentations>=1.3.0
opencv-python-headless>=4.5.5.62
scikit-learn>=1.1.0
Pillow>=9.0.0
numpy>=1.21.0
tqdm>=4.60.0
gdown>=4.4.0
scipy>=1.7.0
torchmetrics>=0.11.0
torchsummary>=1.5.1